requirements.txt
chromadb==0.4.22
langchain==0.1.1
langserve==0.0.39
text-generation==0.6.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==3.16.4
typer==0.9.0
PyMuPDF==1.23.5
sse-starlette==1.6.5
